Add MemberNotFoundException and enhance member deletion with error handling

+++ src/main/java/lk/puLeeNa/LibManagement/controller/MemberController.java
@@ -1,6 +1,7 @@
 package lk.puLeeNa.LibManagement.controller;
 
 import lk.puLeeNa.LibManagement.dto.MemberDTO;
+import lk.puLeeNa.LibManagement.exception.MemberNotFoundException;
 import lk.puLeeNa.LibManagement.service.MemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,13 +20,25 @@ public class MemberController {
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addMember(@RequestBody MemberDTO memberDTO) {
+        if (memberDTO == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         memberService.addMember(memberDTO);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
     @DeleteMapping
     public ResponseEntity<Void> deleteMember(@RequestParam ("memberId") String memberId){
-        memberService.deleteMember(memberId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (memberId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try{
+            memberService.deleteMember(memberId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }catch (MemberNotFoundException e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
     @PatchMapping(value = "/{memberId}", consumes =  MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateMember(@PathVariable String memberId, @RequestBody MemberDTO memberDTO) {
